fix(build): resolve Next.js build errors and type mismatches
- Fix carousel component prop type mismatch
- Convert leaderboard and tournaments pages to client-only to prevent SSR issues
- Add assets.vercel.com to Next.js image remote patterns
- Extract client components to separate files for better organization
